+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 2.0.4:
     * Fix: Display dates in dashboard report chart.
     * Fix: Allow to recalculate taxes for existing order.
+    * Fix: Add missing templates for external and virtual products.
 * 2.0.3 - 2016.11.17:
     * Fix: Product and Order filtering on admin.
     * Fix: Sending stock emails when product has no stock managing.