+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Fix: Allow to recalculate taxes for existing order.
     * Fix: Add missing templates for external and virtual products.
     * Fix: Allow to search product list by SKU.
+    * Fix: Properly calculate tax for saved orders.
 * 2.0.3 - 2016.11.17:
     * Fix: Product and Order filtering on admin.
     * Fix: Sending stock emails when product has no stock managing.